Refactored APIs
