+++ src/main/java/org/cms/events/Event.java
@@ -2,5 +2,5 @@ package org.cms.events;
 
 public enum Event {
 	ASSIGNMENT_UPLOAD,
-	ASSIGNMENT_SOLUTION
+	ASSIGNMENT_SUBMISSION
 }
