+++ src/main/java/org/cms/events/apis/assignment/AssignmentUploadConsumer.java
@@ -1,7 +1,8 @@
-package org.cms.events.apis.assignment.upload;
+package org.cms.events.apis.assignment;
 
 import org.cms.events.apis.Consumer;
 
 public interface AssignmentUploadConsumer extends Consumer {
 	void registerHandler(AssignmentUploadEventHandler handler);
+	void registerFilter(AssignmentUploadEventFilter filter);
 }
