+++ src/main/java/org/cms/events/apis/assignment/AssignmentUploadEventException.java
@@ -1,3 +1,3 @@
-package org.cms.events.apis.assignment.upload;
+package org.cms.events.apis.assignment;
 
 public class AssignmentUploadEventException extends Exception {}
