+++ src/main/java/org/cms/events/apis/assignment/AssignmentUploadEventFilter.java
@@ -1,0 +1,7 @@
+package org.cms.events.apis.assignment;
+
+import org.cms.core.course.Course;
+
+public interface AssignmentUploadEventFilter {
+	boolean satisfies(Course course);
+}
