+++ src/main/java/org/cms/events/apis/assignment/AssignmentUploadEventHandler.java
@@ -1,0 +1,7 @@
+package org.cms.events.apis.assignment;
+
+import org.cms.core.files.assignment.Assignment;
+
+public interface AssignmentUploadEventHandler {
+	void handle(Assignment event);
+}
