+++ src/main/java/org/cms/events/apis/assignment/AssignmentUploadProducer.java
@@ -1,0 +1,8 @@
+package org.cms.events.apis.assignment;
+
+import org.cms.core.files.assignment.Assignment;
+import org.cms.events.apis.Producer;
+
+public interface AssignmentUploadProducer extends Producer {
+	void produce(Assignment event) throws AssignmentUploadEventException;
+}
