--- src/main/java/org/cms/events/apis/assignment/upload/AssignmentUploadEventHandler.java
@@ -1,5 +1,0 @@
-package org.cms.events.apis.assignment.upload;
-
-public interface AssignmentUploadEventHandler {
-	void handle(AssignmentUploadEvent event);
-}
